feat: update database configuration for testing and add new factories
- Changed DB_PORT from 5432 to 5433 in .env.test for testing environment.
- Updated docker-compose.yml to use postgres:17 and added a new postgres_test service for testing.
- Added new dependencies faker and polyfactory to poetry.lock and pyproject.toml for generating test data.
- Removed LoggingMiddleware from main.py.
- Changed scheduled_shift_id type from BigInteger to Integer in shifts.py model.
- Added field validators to remove timezone info from datetime fields in shifts_schemas.py.
- Refactored tests to use AsyncPersistenceFactory for creating test data and added integration tests for shift creation, retrieval, update, and deletion.
- Created new migration script for shifts table.
- Added run_tests.py script to automate test database setup and execution of pytest.

+++ shifts_service/src/models/shifts.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Integer, func
+from sqlalchemy import DateTime, Integer, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.database.base import Base
@@ -11,7 +11,7 @@ class Shift(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     scheduled_shift_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         # ForeignKey("scheduled_shifts.id", ondelete="CASCADE"),
         nullable=False,
     )
